makeTxDbFromUCSC() no longer relies on rtracklayer
makeTxDbFromUCSC() was indirectly relying on rtracklayer via utility
function supportedUCSCtables(). The latter is called internally by
the former to validate the name of the primary table supplied by the
user via the 'tablename' argument.

Until now, supportedUCSCtables() was relying on rtracklayer::trackNames()
and rtracklayer::ucscTables() to do the job, which resulted in a
complicated, inefficient, and unreliable implementation that was hard
to maintain. The new implementation completely bypasses rtracklayer by
querying the UCSC REST API directly at https://api.genome.ucsc.edu. This
results in a much simpler and more efficient implementation which hopefully
will prove more reliable and a lot easier to maintain in the long run
(assuming the UCSC folks don't break it by changing the API of course).

As a consequence, supportedUCSCtables(), makeTxDbFromUCSC(), and
makeTxDbPackageFromUCSC() lose their 'url' argument. Only purpose
of this argument was to point supportedUCSCtables() to the UCSC Genome
Browser mirror to query. However, querying a mirror is no longer an
option now that supportedUCSCtables() uses the UCSC REST API.