Add Taxonomy ID to TxDb object returned by makeTxDbFromEnsembl() (#13)
Like for other makeTxDbFrom* functions, the new 'taxonomyId' argument of makeTxDbFromEnsembl() is set to NA by default. So if the user doesn't supply a Taxonomy ID via the argument, GenomeInfoDb:::lookup_tax_id_by_organism() is used to retrieve the taxonomy ID for the supplied 'organism'. Otherwise GenomeInfoDb:::check_tax_id() is used to check the user-supplied taxonomic ID.